marketplace_ui/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
snowflake-connector-python==4.4.0
